requirements.txt
Django==3.2.5
djangorestframework==3.13.1
numpy==1.21.2
gunicorn==20.1.0
django-cors-headers
whitenoise
drf_yasg
gym
stable_baselines3
